35 - adding mappers

+++ modulith/commodityFinder/src/main/java/io/edpn/backend/modulith/commodityfinder/domain/entity/System.java
@@ -19,7 +19,7 @@ public class System {
     private UUID id;
     private Long eliteId;
     private String name;
-    private Long xCoordinate;
-    private Long yCoordinate;
-    private Long zCoordinate;
+    private Double xCoordinate;
+    private Double yCoordinate;
+    private Double zCoordinate;
 }
